Moved sort helpers into common sort helper file
