+++ src/pages/basic_data/NviAdminNavigationAccordion.tsx
@@ -8,12 +8,12 @@ import { NavigationListAccordion } from '../../components/NavigationListAccordio
 import { NviReportProgressBar } from '../../components/NviReportProgressBar';
 import { LinkCreateButton, NavigationList } from '../../components/PageWithSideMenu';
 import { SelectableButton } from '../../components/SelectableButton';
+import { NviAdminOrderBy } from '../../components/sortSelectors/sortNviTable/nviAdminSortHelpers';
 import { StyledNviStatusBox } from '../../components/styled/Wrappers';
 import { RootState } from '../../redux/store';
 import { dataTestId } from '../../utils/dataTestIds';
 import { getDefaultNviYear } from '../../utils/hooks/useNviCandidatesParams';
 import { UrlPathTemplate } from '../../utils/urlPaths';
-import { NviAdminOrderBy } from './app_admin/nviAdmin/nviAdminHelpers';
 
 export const NviAdminNavigationAccordion = () => {
   const { t } = useTranslation();
